refactor: event dto 패키지 이전
