+++ src/main/java/kr/hhplus/be/server/api/event/service/EventService.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.api.event.service;
 
-import kr.hhplus.be.server.api.event.dto.EventDto;
+import kr.hhplus.be.server.domain.event.dto.EventDto;
 import kr.hhplus.be.server.api.queue.service.QueueService;
 import kr.hhplus.be.server.common.util.CommonUtil;
 import kr.hhplus.be.server.domain.event.repository.EventRepository;
