+++ src/main/java/kr/hhplus/be/server/domain/event/repository/EventQueryRepository.java
@@ -1,8 +1,6 @@
 package kr.hhplus.be.server.domain.event.repository;
 
-import kr.hhplus.be.server.api.event.dto.EventDto;
-import kr.hhplus.be.server.domain.event.entity.Event;
-import org.springframework.data.jpa.repository.JpaRepository;
+import kr.hhplus.be.server.domain.event.dto.EventDto;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
